+++ crates/cheatnet/tests/cheatcodes/spy_events.rs
@@ -483,7 +483,7 @@ fn capture_cairo0_event() {
         ExtendedStateReader {
             dict_state_reader: build_testing_state(),
             fork_state_reader: Some(ForkStateReader::new(
-                "http://188.34.188.184:6060/rpc/v0_6".parse().unwrap(),
+                "http://188.34.188.184:6060/rpc/v0_7".parse().unwrap(),
                 BlockNumber(960_107),
                 temp_dir.path().to_str().unwrap(),
             )),