+++ crates/sncast/src/helpers/constants.rs
@@ -29,7 +29,7 @@ pub const KEYSTORE_PASSWORD_ENV_VAR: &str = "KEYSTORE_PASSWORD";
 pub const CREATE_KEYSTORE_PASSWORD_ENV_VAR: &str = "CREATE_KEYSTORE_PASSWORD";
 
 pub const SCRIPT_LIB_ARTIFACT_NAME: &str = "__sncast_script_lib";
-pub const CONFIG_FILENAME: &str = "snfoundry.toml";
+
 pub const STATE_FILE_VERSION: u8 = 1;
 
 pub const INIT_SCRIPTS_DIR: &str = "scripts";