+++ crates/sncast/tests/e2e/script/deploy.rs
@@ -68,10 +68,11 @@ async fn test_same_salt_and_class_hash_deployed_twice() {
         Unknown location (pc=0:677)
         Unknown location (pc=0:291)
         Unknown location (pc=0:314)
-
         Error in the called contract ([..]):
         Error at pc=0:32:
-        Got an exception while executing a hint: Custom Hint Error: Requested ContractAddress(PatriciaKey(StarkFelt("[..]"))) is unavailable for deployment.
+        Transaction hash = [..]
+        Transaction hash = [..]
+        Got an exception while executing a hint: Requested ContractAddress(PatriciaKey(StarkFelt("[..]"))) is unavailable for deployment.
         Cairo traceback (most recent call last):
         Unknown location (pc=0:174)
         Unknown location (pc=0:127)
@@ -115,10 +116,10 @@ async fn test_invalid_class_hash() {
         Unknown location (pc=0:677)
         Unknown location (pc=0:291)
         Unknown location (pc=0:314)
-
         Error in the called contract ([..]):
         Error at pc=0:32:
-        Got an exception while executing a hint: Custom Hint Error: Class with hash ClassHash(
+        Transaction hash = [..]
+        Got an exception while executing a hint: Class with hash ClassHash(
             StarkFelt(
                 "[..]",
             ),
@@ -166,10 +167,12 @@ async fn test_invalid_call_data() {
         Unknown location (pc=0:677)
         Unknown location (pc=0:291)
         Unknown location (pc=0:314)
-
         Error in the called contract ([..]):
         Error at pc=0:32:
-        Got an exception while executing a hint: Custom Hint Error: Execution failed. Failure reason: [..] ('Failed to deserialize param #2').
+        Got an exception while executing a hint: Execution failed. Failure reason: [..] ('Failed to deserialize param #2').
+        Transaction hash = [..]
+        Error in the called contract ([..]):
+        Execution failed. Failure reason: [..] ('Failed to deserialize param #2').
         Cairo traceback (most recent call last):
         Unknown location (pc=0:174)
         Unknown location (pc=0:127)