+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Forge
 
+#### Added
+
+- variants of cheatcodes with `CheatSpan` (read more [here](https://foundry-rs.github.io/starknet-foundry/testing/using-cheatcodes#setting-cheatcode-span))
+
 #### Fixed
 
 - Events emitted in cairo 0 contracts are now properly collected
@@ -17,6 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 
 - Default `chain_id` has been changed from `SN_GOERLI` to `SN_SEPOLIA`
+- Supported RPC version is now 0.7.0
+
+### Cast
+
+#### Changed
+
+- Supported RPC version is now 0.7.0
 
 
 ## [0.19.0] - 2024-03-06
@@ -27,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-- [`replace_bytecode`](https://foundry-rs.github.io/starknet-foundry/appendix/cheatcodes/replace_bytecode.html)  cheatcode
+- [`replace_bytecode`](https://foundry-rs.github.io/starknet-foundry/appendix/cheatcodes/replace_bytecode.html) cheatcode
 - result of the call to the trace
 - added `--build-profile` flag to the `--test` command. Saves trace data and then builds profiles of test cases which pass and are not fuzz tests. You need [cairo-profiler](https://github.com/software-mansion/cairo-profiler) installed on your system.
 - dependency on the [universal-sierra-compiler](https://github.com/software-mansion/universal-sierra-compiler)